requirements.txt
streamlit
PyPDF2 
spacy 
faiss-cpu 
networkx 
matplotlib 
sentence-transformers
